+++ app/src/client/components/settings/controls/connect.js
@@ -8,16 +8,18 @@ import {
   removeControlsInput,
   resetControls,
   setControlsVisible,
+  setCrosshairVisible,
   bindGamepadToJoypad,
 } from '../../../actions';
 
 const mapStateToProps = state => {
-  const {controls, controlsVisible} = selectSettingsValues(state);
-  return {controls, controlsVisible};
+  const {controls, controlsVisible, crosshairVisible} = selectSettingsValues(state);
+  return {controls, controlsVisible, crosshairVisible};
 };
 
 const mapDispatchToProps = dispatch => ({
   onControlsVisibleChange: visible => dispatch(setControlsVisible(visible)),
+  onCrosshairVisibleChange: visible => dispatch(setCrosshairVisible(visible)),
   onControlsDeviceChange: (port, device) => dispatch(setControlsDevice(port, device)),
   onControlsInputAdd: deviceInput => dispatch(addControlsInput(deviceInput)),
   onControlsInputRemove: sourceInput => dispatch(removeControlsInput(sourceInput)),